Rename BLAS->blas and CL->opencl
