+++ src/neural/network_blas.cc
@@ -18,8 +18,8 @@
 
 #include "neural/network.h"
 #include "neural/factory.h"
-#include "neural/BLAS/blas.h"
-#include "neural/BLAS/transforms.h"
+#include "neural/blas/blas.h"
+#include "neural/blas/transforms.h"
 
 #include <algorithm>
 #include <cassert>
