+++ src/neural/opencl/clblast_level3/common.opencl
