+++ src/neural/opencl/clsource/config.opencl
