+++ src/neural/opencl/clsource/convolve1.opencl
