+++ src/neural/opencl/clsource/convolve3.opencl
